+++ VM-Core/src/main/java/com/validation/manager/core/db/controller/UserStatusJpaController.java
@@ -12,6 +12,7 @@ import javax.persistence.EntityNotFoundException;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
 import com.validation.manager.core.db.VmUser;
+import com.validation.manager.core.db.controller.exceptions.IllegalOrphanException;
 import com.validation.manager.core.db.controller.exceptions.NonexistentEntityException;
 import java.util.ArrayList;
 import java.util.List;
@@ -65,7 +66,7 @@ public class UserStatusJpaController implements Serializable {
         }
     }
 
-    public void edit(UserStatus userStatus) throws NonexistentEntityException, Exception {
+    public void edit(UserStatus userStatus) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -73,6 +74,18 @@ public class UserStatusJpaController implements Serializable {
             UserStatus persistentUserStatus = em.find(UserStatus.class, userStatus.getId());
             List<VmUser> vmUserListOld = persistentUserStatus.getVmUserList();
             List<VmUser> vmUserListNew = userStatus.getVmUserList();
+            List<String> illegalOrphanMessages = null;
+            for (VmUser vmUserListOldVmUser : vmUserListOld) {
+                if (!vmUserListNew.contains(vmUserListOldVmUser)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain VmUser " + vmUserListOldVmUser + " since its userStatusId field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
             List<VmUser> attachedVmUserListNew = new ArrayList<VmUser>();
             for (VmUser vmUserListNewVmUserToAttach : vmUserListNew) {
                 vmUserListNewVmUserToAttach = em.getReference(vmUserListNewVmUserToAttach.getClass(), vmUserListNewVmUserToAttach.getId());
@@ -81,12 +94,6 @@ public class UserStatusJpaController implements Serializable {
             vmUserListNew = attachedVmUserListNew;
             userStatus.setVmUserList(vmUserListNew);
             userStatus = em.merge(userStatus);
-            for (VmUser vmUserListOldVmUser : vmUserListOld) {
-                if (!vmUserListNew.contains(vmUserListOldVmUser)) {
-                    vmUserListOldVmUser.setUserStatusId(null);
-                    vmUserListOldVmUser = em.merge(vmUserListOldVmUser);
-                }
-            }
             for (VmUser vmUserListNewVmUser : vmUserListNew) {
                 if (!vmUserListOld.contains(vmUserListNewVmUser)) {
                     UserStatus oldUserStatusIdOfVmUserListNewVmUser = vmUserListNewVmUser.getUserStatusId();
@@ -115,7 +122,7 @@ public class UserStatusJpaController implements Serializable {
         }
     }
 
-    public void destroy(Integer id) throws NonexistentEntityException {
+    public void destroy(Integer id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -127,10 +134,16 @@ public class UserStatusJpaController implements Serializable {
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The userStatus with id " + id + " no longer exists.", enfe);
             }
-            List<VmUser> vmUserList = userStatus.getVmUserList();
-            for (VmUser vmUserListVmUser : vmUserList) {
-                vmUserListVmUser.setUserStatusId(null);
-                vmUserListVmUser = em.merge(vmUserListVmUser);
+            List<String> illegalOrphanMessages = null;
+            List<VmUser> vmUserListOrphanCheck = userStatus.getVmUserList();
+            for (VmUser vmUserListOrphanCheckVmUser : vmUserListOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This UserStatus (" + userStatus + ") cannot be destroyed since the VmUser " + vmUserListOrphanCheckVmUser + " in its vmUserList field has a non-nullable userStatusId field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(userStatus);
             em.getTransaction().commit();
